Add intersection in model space to SceneRayCastResults
Use this to compute barycentric coordinates for triangle - since this
is also in model space. This fixes the bug of getting strange
barycentric coordinates not inside the triangle.

We also use this to compute distance to geometry. This should fix
incorrect z ordering between models with different scales in their
model matrix.