Remove all the immediate mode rendering holdovers
Everything are now meshes. This greatly simplifies the backends, since
they can perform whatever pre-processing they want on the data to
create their own "mesh" types, so we don't need to support a path for
doing that computation on the fly during rendering (as was required
for immediate mode rendering). The mesh creation interface is now
painless enough that removing immediate mode rendering is no big
deal.